:recycle: [open-formulieren/open-forms#4917] Refactor so that FormNavigation is used everywhere
Rather than only using this for step navigation and constructing a
similar composite on the summary pages, we now apply the same
FormNavigation component but with different configurations and/or
submit buttons for their appropriate context.

This actually leads to nicer code, since a number of props were just
being passed down to pass them to the button component, but we can
now provide those directly to the component and pollute the container
component less.
