+++ src/components/FormNavigation/index.ts
@@ -1,0 +1,5 @@
+import FormNavigation from './FormNavigation';
+
+export {default as FormSubmitButton} from './FormSubmitButton';
+export {default as StepSubmitButton} from './StepSubmitButton';
+export default FormNavigation;
