+++ src/components/FormNavigation/test.spec.tsx
@@ -6,7 +6,7 @@ import {MemoryRouter} from 'react-router';
 import {LiteralsProvider} from 'components/Literal';
 import {SUBMISSION_ALLOWED} from 'components/constants';
 
-import FormNavigation from './index';
+import FormNavigation, {StepSubmitButton} from './index';
 
 const LITERALS = {
   nextText: {value: '', resolved: 'Next step'},
@@ -14,7 +14,7 @@ const LITERALS = {
   previousText: {value: '', resolved: 'Previous step'},
 };
 
-const Wrap = ({children}) => (
+const Wrap: React.FC<React.PropsWithChildren> = ({children}) => (
   <IntlProvider locale="nl" messages={messagesNL}>
     <MemoryRouter>
       <LiteralsProvider literals={LITERALS}>{children}</LiteralsProvider>
@@ -28,13 +28,16 @@ it('Last step of submittable form, button is present', () => {
   render(
     <Wrap>
       <FormNavigation
-        canSubmitStep={true}
-        canSubmitForm={SUBMISSION_ALLOWED.yes}
+        submitButton={
+          <StepSubmitButton
+            canSubmitForm={SUBMISSION_ALLOWED.yes}
+            canSubmitStep
+            isLastStep
+            isCheckingLogic={false}
+          />
+        }
         canSuspendForm={true}
         isAuthenticated={false}
-        isLastStep={true}
-        isCheckingLogic={false}
-        loginRequired={false}
         onFormSave={mockFunction}
         onDestroySession={mockFunction}
         previousPage="#"
@@ -54,13 +57,16 @@ it('Last step of non-submittable form with overview, button is present', () => {
   render(
     <Wrap>
       <FormNavigation
-        canSubmitStep={true}
-        canSubmitForm={SUBMISSION_ALLOWED.noWithOverview}
+        submitButton={
+          <StepSubmitButton
+            canSubmitForm={SUBMISSION_ALLOWED.noWithOverview}
+            canSubmitStep
+            isLastStep
+            isCheckingLogic={false}
+          />
+        }
         canSuspendForm={true}
         isAuthenticated={false}
-        isLastStep={true}
-        isCheckingLogic={false}
-        loginRequired={false}
         previousPage="#"
         onFormSave={mockFunction}
         onDestroySession={mockFunction}
@@ -80,13 +86,16 @@ it('Last step of non-submittable form without overview, button is NOT present', 
   render(
     <Wrap>
       <FormNavigation
-        canSubmitStep={true}
-        canSubmitForm={SUBMISSION_ALLOWED.noWithoutOverview}
+        submitButton={
+          <StepSubmitButton
+            canSubmitForm={SUBMISSION_ALLOWED.noWithoutOverview}
+            canSubmitStep
+            isLastStep
+            isCheckingLogic={false}
+          />
+        }
         canSuspendForm={true}
         isAuthenticated={false}
-        isLastStep={true}
-        isCheckingLogic={false}
-        loginRequired={false}
         previousPage="#"
         onFormSave={mockFunction}
         onDestroySession={mockFunction}
@@ -106,13 +115,16 @@ it('Non-last step of non-submittable form without overview, button IS present', 
   render(
     <Wrap>
       <FormNavigation
-        canSubmitStep={true}
-        canSubmitForm={SUBMISSION_ALLOWED.noWithoutOverview}
+        submitButton={
+          <StepSubmitButton
+            canSubmitForm={SUBMISSION_ALLOWED.noWithOverview}
+            canSubmitStep
+            isLastStep={false}
+            isCheckingLogic={false}
+          />
+        }
         canSuspendForm={true}
         isAuthenticated={false}
-        isLastStep={false}
-        isCheckingLogic={false}
-        loginRequired={false}
         previousPage="#"
         onFormSave={mockFunction}
         onDestroySession={mockFunction}
@@ -132,13 +144,16 @@ it('Suspending form allowed, button is present', () => {
   render(
     <Wrap>
       <FormNavigation
-        canSubmitStep={true}
-        canSubmitForm={SUBMISSION_ALLOWED.yes}
+        submitButton={
+          <StepSubmitButton
+            canSubmitForm={SUBMISSION_ALLOWED.yes}
+            canSubmitStep
+            isLastStep
+            isCheckingLogic={false}
+          />
+        }
         canSuspendForm={true}
         isAuthenticated={false}
-        isLastStep={true}
-        isCheckingLogic={false}
-        loginRequired={false}
         previousPage="#"
         onFormSave={mockFunction}
         onDestroySession={mockFunction}
@@ -155,13 +170,16 @@ it('Suspending form not allowed, button is NOT present', () => {
   render(
     <Wrap>
       <FormNavigation
-        canSubmitStep={true}
-        canSubmitForm={SUBMISSION_ALLOWED.yes}
+        submitButton={
+          <StepSubmitButton
+            canSubmitForm={SUBMISSION_ALLOWED.yes}
+            canSubmitStep
+            isLastStep
+            isCheckingLogic={false}
+          />
+        }
         canSuspendForm={false}
         isAuthenticated={false}
-        isLastStep={true}
-        isCheckingLogic={false}
-        loginRequired={false}
         previousPage="#"
         onFormSave={mockFunction}
         onDestroySession={mockFunction}
