+++ src/components/appointments/SubmitRow.jsx
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import {useNavigate} from 'react-router';
 
-import FormNavigation from 'components/FormNavigation';
+import FormNavigation, {StepSubmitButton} from 'components/FormNavigation/index';
 import {LiteralsProvider} from 'components/Literal';
 import {SUBMISSION_ALLOWED} from 'components/constants';
 
@@ -16,18 +16,21 @@ const SubmitRow = ({canSubmit, nextText, previousText = '', navigateBackTo = ''}
       }}
     >
       <FormNavigation
-        canSubmitStep={canSubmit}
-        canSubmitForm={SUBMISSION_ALLOWED.yes}
+        submitButton={
+          <StepSubmitButton
+            canSubmitStep={canSubmit}
+            isCheckingLogic={false}
+            canSubmitForm={SUBMISSION_ALLOWED.yes}
+            isLastStep={false}
+          />
+        }
         canSuspendForm={false}
-        isAuthenticated={false} // TODO -> if authenticated (for prefill), logout must be shown
-        isLastStep={false}
-        isCheckingLogic={false}
-        loginRequired={false}
-        hideAbortButton
-        previousPage={navigateBackTo ? `../${navigateBackTo}` : ''}
         onFormSave={() => {}}
-        onNavigatePrevPage={navigateBackTo ? () => navigate(`../${navigateBackTo}`) : undefined}
+        hideAbortButton
+        isAuthenticated={false} // TODO -> if authenticated (for prefill), logout must be shown
         onDestroySession={() => {}}
+        previousPage={navigateBackTo ? `../${navigateBackTo}` : ''}
+        onNavigatePrevPage={navigateBackTo ? () => navigate(`../${navigateBackTo}`) : undefined}
       />
     </LiteralsProvider>
   );
